chore: Update CSS for table of contents
This commit updates the custom.css file to modify the styling of the table of contents (TOC) in the documentation. It adjusts the padding of the TOC drawer and adds a new CSS class for the TOC tree container to enable horizontal scrolling and prevent wrapping of the content.

The changes also include adding a new CSS class for the edit this page link in the content icon container, allowing it to be displayed flexibly with aligned icons.

These modifications improve the readability and usability of the documentation.